pyproject.toml
[project]
name = "Flask-Caching"
version = "2.3.1"
description = "Adds caching support to Flask applications."
readme = "README.md"
license = "BSD-3-Clause"
license-files = ["LICENSE"]
authors = [{ name = "Peter Justin", email="peter.justin@outlook.com" }]
maintainers = [{ name = "Pallets Ecosystem", email = "contact@palletsprojects.com" }]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Environment :: Web Environment',
    'Intended Audience :: Developers',
    'Operating System :: OS Independent',
    'Programming Language :: Python',
]
requires-python = ">=3.10"
dependencies = [
    "flask>=2.0",
    "cachelib>=0.9.0",
]

[project.urls]
Documentation = "https://flask-caching.readthedocs.io"
Changes = "https://github.com/pallets-eco/flask-caching/releases/"
Source = "https://github.com/pallets-eco/flask-caching"
Chat = "https://discord.gg/pallets"

[dependency-groups]
dev = [
    "ruff",
    "tox>=4.39",
    "tox-uv",
    {include-group = "pre-commit"},
    {include-group = "docs"},
    {include-group = "tests"},
    {include-group = "typing"},
]
pre-commit = [
    "pre-commit",
    "pre-commit-uv",
]
docs = [
    "pallets-sphinx-themes",
    "sphinx",
    "sphinx-issues",
    "sphinx-tabs",
    "sphinx-intl>=2.3.1",
    "sphinxcontrib-log-cabinet",
]
tests = [
    "flask[async]",
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "pytest-xprocess",
    "libmc",
    "redis",
    "asgiref",
]
typing = [
    "mypy",
    "pyright",
    "pytest",
]
gha-update = [
    "gha-update ; python_full_version >= '3.12'",
]

[build-system]
requires = ["flit_core>=3.11,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "flask_caching"

[tool.flit.sdist]
include = [
    "docs/",
    "examples/",
    "tests/",
    "CHANGES.rst",
    "uv.lock"
]
exclude = [
    "docs/_build/",
]

[tool.uv]
default-groups = ["dev", "pre-commit", "tests", "typing"]

[tool.ruff]
src = ["src"]
fix = true
show-fixes = true
output-format = "full"

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # pyflakes
    "I",  # isort
    "UP",  # pyupgrade
    "W",  # pycodestyle warning
]
ignore = [
    "UP007",  # Use X | Y for type annotations; doesn't work with some of our (stringy) type definitions - should aim to resolve this
]

[tool.ruff.lint.isort]
force-single-line = true
order-by-type = false

[tool.coverage.run]
branch = true
source = ["flask_caching", "tests"]
omit = ["src/flask_caching/contrib/*"]

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.report]
exclude_also = [
    "if t.TYPE_CHECKING",
    "raise NotImplementedError",
    ": \\.{3}",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "error",
    "default::pytest.PytestUnraisableExceptionWarning",
]

[tool.mypy]
python_version = "3.10"
files = ["src/flask_caching/"]
warn_unused_configs = true

# TODO if all of these can be enabled we can just set 'strict = true'
disallow_subclassing_any = true
# disallow_untyped_calls = true
# disallow_untyped_defs = true
# disallow_incomplete_defs = true
# check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
# warn_return_any = true
# no_implicit_reexport = true
strict_equality = true
strict_concatenate = true

[tool.pyright]
pythonVersion = "3.10"
include = ["src"]
typeCheckingMode = "basic"


[tool.tox]
env_list = [
    "py3.14",
    "py3.13", "py3.12", "py3.11", "py3.10",
    "style",
    "typing",
    "docs",
]

[tool.tox.env_run_base]
description = "pytest on latest dependency versions"
runner = "uv-venv-lock-runner"
package = "wheel"
wheel_build_env = ".pkg"
constrain_package_deps = true
use_frozen_constraints = true
dependency_groups = ["tests"]
env_tmp_dir = "{toxworkdir}/tmp/{envname}"
commands = [[
    "pytest", "-v", "--tb=short", "--basetemp={env_tmp_dir}",
    {replace = "posargs", default = [], extend = true},
]]

[tool.tox.env.style]
description = "run all pre-commit hooks on all files"
dependency_groups = ["pre-commit"]
skip_install = true
commands = [["pre-commit", "run", "--all-files"]]

[tool.tox.env.typing]
description = "run static type checkers"
dependency_groups = ["typing"]
commands = [
    ["mypy"],
    ["pyright"],
]

[tool.tox.env.docs]
description = "build docs"
dependency_groups = ["docs"]
commands = [["sphinx-build", "-E", "-W", "-b", "html", "docs", "docs/_build/"]]

[tool.tox.env.update-actions]
description = "update GitHub Actions pins"
labels = ["update"]
dependency_groups = ["gha-update"]
skip_install = true
commands = [["gha-update"]]

[tool.tox.env.update-pre_commit]
description = "update pre-commit pins"
labels = ["update"]
dependency_groups = ["pre-commit"]
skip_install = true
commands = [["pre-commit", "autoupdate", "--freeze", "-j4"]]

[tool.tox.env.update-requirements]
description = "update uv lock"
labels = ["update"]
dependency_groups = []
no_default_groups = true
skip_install = true
commands = [["uv", "lock", {replace = "posargs", default = ["-U"], extend = true}]]
